Mutualise extern functions
