+++ src/components/structures/UserSettings.js
@@ -36,6 +36,7 @@ const SdkConfig = require('../../SdkConfig');
 import Analytics from '../../Analytics';
 import AccessibleButton from '../views/elements/AccessibleButton';
 import { _t, _td } from '../../languageHandler';
+import { isCurrentUserExtern } from '../../Users'
 import * as languageHandler from '../../languageHandler';
 import * as FormattingUtils from '../../utils/FormattingUtils';
 
@@ -973,12 +974,6 @@ module.exports = React.createClass({
         CallMediaHandler.setVideoInput(deviceId);
     },
 
-    _isUserExtern: function() {
-        const hsUrl = MatrixClientPeg.get().getHomeserverUrl();
-        return hsUrl.includes('.e.') || hsUrl.includes('.externe.');
-    },
-
-
     _requestMediaPermissions: function(event) {
         const getUserMedia = (
             window.navigator.getUserMedia || window.navigator.webkitGetUserMedia || window.navigator.mozGetUserMedia
@@ -1165,7 +1160,7 @@ module.exports = React.createClass({
             this.state.avatarUrl ? MatrixClientPeg.get().mxcUrlToHttp(this.state.avatarUrl) : null
         );
 
-        let isUserExternam = this._isUserExtern();
+        let isUserExtern = isCurrentUserExtern();
 
         const threepidsSection = this.state.threepids.map((val, pidIndex) => {
             const id = "3pid-" + val.address;
@@ -1215,7 +1210,7 @@ module.exports = React.createClass({
 
         let redlistOptionSection;
 
-        if (!isUserExternam) {
+        if (!isUserExtern) {
             redlistOptionSection = (
                 <div className="mx_UserSettings_redList">
                     <input id="redlistOption"
